aineon-enterprise/requirements.txt
# AINEON Flash Loan Engine - Production Requirements
# Python 3.10+ compatible dependencies

# Core Web3 and Blockchain
web3==6.15.1
websockets==12.0
eth-account==0.11.0
eth-brownie==1.20.4
pycoingecko==3.1.0

# AsyncIO and HTTP
aiohttp==3.9.1
aiofiles==23.2.1
httpx==0.25.2

# Data Processing - Python 3.10+ compatible versions
numpy==1.24.4
pandas==2.1.4
scipy==1.11.4

# Configuration and Environment
python-dotenv==1.0.0
pydantic==2.5.0
pyyaml==6.0.1

# Logging and Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Database (optional)
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Security
cryptography==41.0.8
pyjwt==2.8.0

# Network and API
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
flask==3.0.0
flask-cors==4.0.0

# Performance
uvloop==0.19.0
orjson==3.9.10

# Caching
cachetools==5.3.2
diskcache==5.6.3

# Error Handling
rich==13.7.0
typer==0.9.0

# Utility Libraries
click==8.1.7
tqdm==4.66.1
colorama==0.4.6

# Mathematical Operations
sympy==1.12

# Machine Learning
scikit-learn==1.3.2

# Date and Time
python-dateutil==2.8.2
pytz==2023.3

# Cryptocurrency Specific
ccxt==4.1.65

# Blockchain Specific
eth-abi==4.1.0
eth-utils==3.0.0
aiohttp-cors==0.7.0

